test/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses-json==0.5.14
distro==1.9.0
exceptiongroup==1.2.1
faiss-cpu==1.8.0
filelock==3.15.1
frozenlist==1.4.1
fsspec==2024.6.0
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.1
google-search-results==2.4.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.181
langchain-core==0.2.6
langchain-text-splitters==0.2.1
langsmith==0.1.77
lazy_loader==0.4
llama-index==0.6.12
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
openapi-schema-pydantic==1.2.4
opencv-python-headless==4.5.4.60
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pydantic==1.10.16
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.24
sympy==1.12.1
tenacity==8.3.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing-inspect==0.8.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
yarl==1.9.4
youtube-transcript-api==0.6.2
